CardsetChildren 컴포넌트로 분리하기
