+++ src/components/Cardset.jsx
@@ -1,9 +1,10 @@
 import { Link } from 'react-router-dom';
 
-import { BsCheckLg } from 'react-icons/bs';
-import { FaFolder, FaPen, FaPlayCircle } from 'react-icons/fa';
-
 import styled from '@emotion/styled';
+
+import { FaPen, FaPlayCircle } from 'react-icons/fa';
+
+import CardsetChildren from './CardsetChildren';
 
 const Wrapper = styled.div({
   width: '100%',
@@ -26,34 +27,6 @@ const SubTitle = styled.h2({
   marginBottom: '10px',
   color: '#686868',
   borderBottom: '1px solid #C6C6C6',
-});
-
-const CardsetBoxField = styled.li({
-  width: '90%',
-  fontWeight: 'bold',
-  marginBottom: '10px',
-  backgroundColor: '#F3F3F3',
-  padding: '5px 0px 5px 5px',
-  display: 'flex',
-  justifyContent: 'space-between',
-  alignItems: 'center',
-  '& a:hover': {
-    textDecoration: 'underline',
-  },
-});
-
-const CardBoxField = styled.li({
-  width: '90%',
-  fontWeight: 'bold',
-  marginBottom: '10px',
-  backgroundColor: '#FAFAFA',
-  padding: '5px 0px 5px 5px',
-  display: 'flex',
-  alignItems: 'center',
-});
-
-const CardsetBoxTitle = styled.div({
-  display: 'flex',
 });
 
 const CardsetInfo = styled.div({
@@ -111,50 +84,9 @@ export default function Cardset({ cardsetInfo, cardsetChildren }) {
       <SubTitle>
         Cards
       </SubTitle>
-      <ul>
-        {cardsetChildren.map((child) => {
-          if (child.type === 'CARDSET') {
-            return (
-              <CardsetBoxField
-                key={child.id}
-              >
-                <CardsetBoxTitle>
-                  <IconBox>
-                    <FaFolder />
-                  </IconBox>
-                  <Link to={`/cardsets/${child.id}`}>
-                    {child.title}
-                  </Link>
-                </CardsetBoxTitle>
-                <div>
-                  <IconBox
-                    type="button"
-                  >
-                    <Link to={`/studio/${child.id}`}>
-                      <FaPen />
-                    </Link>
-                  </IconBox>
-                  <IconBox
-                    type="button"
-                  >
-                    <FaPlayCircle />
-                  </IconBox>
-                </div>
-              </CardsetBoxField>
-            );
-          }
-          return (
-            <CardBoxField
-              key={child.id}
-            >
-              <IconBox>
-                <BsCheckLg />
-              </IconBox>
-              {child.title}
-            </CardBoxField>
-          );
-        })}
-      </ul>
+      <CardsetChildren
+        cardsetChildren={cardsetChildren}
+      />
     </Wrapper>
   );
 }
